Split out team component
